Publish submodules from pacakge root
