+++ .github/workflows/npm-publish.yml
@@ -26,6 +26,6 @@ jobs:
           node-version: 12
           registry-url: https://registry.npmjs.org/
       - run: yarn
-      - run: npm publish
+      - run: yarn publish-stable
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
