+++ package.json
@@ -2,8 +2,8 @@
   "name": "react-helsinki-headless-cms",
   "version": "0.0.0",
   "description": "React components for displaying Headless CMS content according to guidelines set by HDS",
-  "main": "lib/index.js",
-  "module": "lib-esm/index.js",
+  "main": "cjs/index.js",
+  "module": "index.js",
   "scripts": {
     "dev": "yarn storybook",
     "lint": "yarn eslint \"src/**/*.{ts,tsx}\" && yarn prettier ./src -c",
@@ -12,6 +12,7 @@
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "publish-canary": "node ./scripts/publish-canary.js",
+    "publish-stable": "node ./scripts/publish-stable.js",
     "generate:graphql": "DOTENV_CONFIG_PATH=./.env graphql-codegen -r dotenv/config --config codegen.yml"
   },
   "repository": "https://github.com/City-of-Helsinki/react-helsinki-headless-cms.git",
@@ -21,10 +22,10 @@
   },
   "homepage": "https://github.com/City-of-Helsinki/react-helsinki-headless-cms#readme",
   "peerDependencies": {
+    "@apollo/client": "^3.5.10",
     "hds-react": ">=1.11.1",
     "react": ">=16.8.0",
-    "react-dom": ">=16.8.0",
-    "@apollo/client": "^3.5.10"
+    "react-dom": ">=16.8.0"
   },
   "peerDependenciesMeta": {
     "@apollo/client": {
@@ -85,6 +86,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "rollup": "^2.70.0",
+    "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-includepaths": "^0.2.4",
     "rollup-plugin-polyfill-node": "^0.8.0",
     "rollup-plugin-postcss": "^4.0.2",
